requirements.txt
flask
requests
duckdb
pandas
rank_bm25
langchain_community
langchain-openai
langchain_huggingface
faiss-gpu